Extended schema to support diagrammatic data
This move required changing the root element of our objects, which is now called a Project. A project contains a model and diagrams. A diagram contains shapes, lines, and labels, all of which refer to elements in the model (e.g. classes, relations, generalization sets). I managed to define a single schema for the these three elements types that are shown in diagrams. All previous tests have been adjusted to account for this new structure.